refactor : 가독성을 위한 코드 스타일 변경
