+++ api/src/main/java/org/meotppo/webti/log/aop/LoggingAspect.java
@@ -1,8 +1,7 @@
 package org.meotppo.webti.log.aop;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import java.io.PrintWriter;
-import java.io.StringWriter;
+import java.io.IOException;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -11,33 +10,34 @@ import org.aspectj.lang.annotation.Around;
 import org.aspectj.lang.annotation.Aspect;
 import org.springframework.stereotype.Component;
 
-@Aspect
 @Slf4j
+@Aspect
 @Component
 @RequiredArgsConstructor
 public class LoggingAspect {
 
     private final ObjectMapper objectMapper;
 
-    @Around("within(*..*Controller) && !@annotation(org.meotppo.webti.log.aop.NotLogging)")
+    @Around("within(@org.springframework.web.bind.annotation.RestController *) && "
+            + "!@annotation(org.meotppo.webti.log.aop.NotLogging)")
     public Object requestLogging(ProceedingJoinPoint joinPoint) throws Throwable {
-        final LogInfoDto logInfoDto = new LogInfoDto(joinPoint, joinPoint.getTarget().getClass(), objectMapper);
-
+        LogInfoDto logInfoDto = new LogInfoDto(joinPoint, joinPoint.getTarget().getClass(), objectMapper);
         try {
-            final Object result = joinPoint.proceed(joinPoint.getArgs());
-            final String logMessage = objectMapper.writeValueAsString(Map.entry("logInfo", logInfoDto));
+            Object result = joinPoint.proceed(joinPoint.getArgs());
+            String logMessage = convertToJson(Map.entry("logInfo", logInfoDto));
 
             log.info(logMessage);
             return result;
         } catch (Exception e) {
-            final StringWriter sw = new StringWriter();
-            e.printStackTrace(new PrintWriter(sw));
-            final String exceptionAsString = sw.toString();
-            logInfoDto.setException(exceptionAsString);
-            final String logMessage = objectMapper.writeValueAsString(logInfoDto);
+            logInfoDto.setException(e.toString());
+            String logMessage = convertToJson(logInfoDto);
 
             log.error(logMessage);
             throw e;
         }
+    }
+
+    private String convertToJson(Object object) throws IOException {
+        return objectMapper.writeValueAsString(object);
     }
 }
